LayoutTests/fast/dom/HTMLElement/attr-dir-auto-change-child-node.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
<script src="../../js/resources/js-test-pre.js"></script>
<style>
.testDiv {-webkit-border-start: 5px solid green; -webkit-border-end: 5px solid red; }
</style>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>

<div id="test" dir="auto" class="testDiv"></div>
<script>

description('Test that directionality is re-evaluated when an element with dir=auto set, has it\'s first child modified.');

var el = document.getElementById("test");
var child = document.createElement("div");
child.innerHTML="test";
el.appendChild(child);
shouldBe("document.defaultView.getComputedStyle(el, null).getPropertyValue('border-right-color')", "'rgb(255, 0, 0)'");

var child2 = document.createElement("div");
child2.innerHTML="מקור השם עברית";
el.insertBefore(child2, child);
shouldBe("document.defaultView.getComputedStyle(el, null).getPropertyValue('border-right-color')", "'rgb(0, 128, 0)'");
el.style.display = "none";
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
